+++ app/src/main/java/com/example/my_shoplist_application/domain/models/Ingredients.kt
@@ -3,7 +3,7 @@ package com.example.my_shoplist_application.domain.models
 data class Ingredients(
     val id: Int = 0,
     val ingredientName: String,
-    val ingredientQuantity: Float,
+    val ingredientQuantity: Int,
     val ingredientUnit: MeasurementUnit,
     val shopListId: Int = 0,
     var isBought: Boolean = false
